import drupal-6.x-notify-1.2:
The notify module allows users to subscribe to periodic emails which
include all new or revised content and/or comments much like the daily
news letters sent by some websites.
Even if this feature is not configured for normal site users, it can
be a useful feature for an administrator of a site to receive
notification of new content submissions and comment posts.