import mspdebug 0.10
MSPDebug is a free debugger for use with MSP430 MCUs. It supports
FET430UIF, eZ430, RF2500 and Olimex MSP-JTAG-TINY programmers. It can be
used as a proxy for gdb or as an independent debugger with support for
programming, disassembly and reverse engineering.

With a patch from jakemsr@ to work around a ugen/libusb issue where
usb_bulk_read return 0 on timeout, which isn't deal with properly.

ok ajacoutot@