Import ruby-liquid 2.2.2, dependency of ruby-jekyll
Liquid is a template engine which I wrote for very specific
requirements:

* It has to have beautiful and simple markup. Template engines which
don't produce good looking markup are no fun to use.

* It needs to be non evaling and secure. Liquid templates are made so
that users can edit them. You don't want to run code on your server
which your users wrote.

* It has to be stateless. Compile and render steps have to be seperate
so that the expensive parsing and compiling can be done once and later
on you can just render it passing in a hash with local variables and
objects.