Import tumbler 0.1.6
Tumbler is a D-Bus service for applications to request thumbnails for
various URI schemes and MIME types. It is an implementation of the
thumbnail management D-Bus specification described on
http://live.gnome.org/ThumbnailerSpec written in an object-oriented
fashion using GLib and GObject.

thunar can make use of it but it's not required by it (as is
hunar-thumbnailers package..)