import tidyp 1.04 (which is a forked continuation for tidy, ports will be migrated eventually)
tidyp is a program that can validate your HTML, as well as modify it to
be more clean and standard. tidyp does not validate HTML 5.
libtidyp is the library on which the program is based. It can be used by
any other program that can interface to it.

ok sthen@