Depend on devel/hs-mtl (mtl-2) so it doesn't pick up mtl-1 by accident (if ghc is installed but hs-mtl isn't).
Rule of thumb: if some hs-* port works with both mtl-1 and
mtl-2, let it depend on devel/hs-mtl. If it strictly requires
mtl-1, then don't let it depend on devel/hs-mtl (that part is
obvious). If it doesn't specify the mtl version i the .cabal
file and breaks with devel/hs-mtl installed, patch the .cabal
file.

I'll rebuild and check all the haskell junk on my machine and
will see what can be done to hs-* ports that don't build with
mtl-2 (devel/hs-mtl). If they can be fixed to work with mtl-2,
mtl-1 (in lang/ghc) will be removed.

What a mess!