Import xfce4-power-manager 1.0.10:
This software is a power manager for the Xfce desktop, and relies on
UPower to gather the power sources on the computer. In addition,
xfce4-power-manager provides a set of freedesktop-compliant DBus
interfaces to inform other applications about current power level so
that they can adjust their power consumption.

ok ajacoutot@