import ports/sysutils/gsmartcontrol, incorporating bits from giovanni@'s version of the same port, tweak + ok aja@
GSmartControl is a graphical user interface for smartctl (from the
smartmontools package), which is a tool for querying and controlling
SMART (Self-Monitoring, Analysis, and Reporting Technology) data
on modern hard disk drives.

It allows you to inspect the drive's SMART data to determine its
health, as well as run various tests on it.