import gnome-contacts 3.2.0.1
GNOME Contacts will pull together your contacts from your different
online accounts and integrate them into a single address book. This
means that looking up someone in GNOME Contacts will give you access to
a summary of all your online relationships with that person (or as many
as we are legally allowed to), including the various ways you can
communicate with that person and links to the different places where
they exist on the web.

ok aja@