Import icedtea-web 1.1.2.
The IcedTea-Web project provides a Free Software web browser plugin
running applets written in the Java programming language and an
implementation of Java Web Start, originally based on the NetX project.
It works in gecko-based AND webkit-based browsers, and is meant as a
replacement for the dead/unmaintained/dreaded sun java plugin.

It is still a it rough at the edges (java process sometimes doesn't
finish when tab is closed, awt exceptions upon first paint), but let's
import it so that people can try it and debug it.

ok sthen@