Import chromium-bsu-0.9.15.
You are captain of the cargo ship Chromium B.S.U., responsible for
delivering supplies to our troops on the front line. Your ship has a
small fleet of robotic fighters which you control from the relative
safety of the Chromium vessel.

from Antti Harri (maintainer) with some tweaks
committing on behalf of and ok landry@