Import libsecret-0.1.
libsecret is a library for storing and retrieving passwords and other
secrets. It communicates with the "Secret Service" using DBus.
gnome-keyring and ksecretservice are both implementations of a Secret
Service.

It is a replacement for libgnome-keyring. Nothing will use it as yet,
but it's better to have it in-tree now to start fixing portability
issues with upstream. It will also ease our work on the next major
release of GNOME.