import p5-DBIx-Class-IntrospectableM2M from Andrew Fresh, ok landry@
Because the many-to-many relationships are not real relationships,
they can not be introspected with DBIx::Class. Many-to-many
relationships are actually just a collection of convenience methods
installed to bridge two relationships. This DBIx::Class component
can be used to store all relevant information about these
non-relationships so they can later be introspected and examined.