import p5-Proc-PID-File
s module is useful for writers of daemons and other processes that
need to tell whether they are already running, in order to prevent
multiple process instances. The module accomplishes this via *nix-style
pidfiles, which are files that store a process identifier.

tested on i386 and macppc, merdely on amd64

OK merdely