fix caching issues.
- during the scanning stage, we can rely on more than sizes. Specifically,
for files with cached sha values: detect problems early, zap the files, so
the new ones do fetch.

- do not allow "negative" caching: if the cached file doesn't match, just
run the checksum again to make sure (manual download would tamper with that).
This should allow builders to forget about the existence of
/usr/ports/distfiles/distinfo again.

- remove bad files so that fetch has a chance to work (todo: log some more
info, yeah landry...)

- zap code from (checksum) proper  that's no longer in-use.

okay jasper@

(gets in because fixing the mirrors for the release is important, and dpb -F
would not do the right thing without manual intervention).