Initial import of p5-Gravatar-URL 1.06
A Gravatar is a Globally Recognized Avatar for a given email address.
This allows you to have a global picture associated with your email
address.  You can look up the Gravatar for any email address by
constructing a URL to get the image from gravatar.com.  This module
does that.

OK sthen@