Import 4 new simple QT games from gottcode.org.
Hexalate is a color matching game. The goal of the game is to rotate and
position the circles so that each touching line matches in color. You
rotate circles by right clicking, and you move circles by dragging them.

Simsu is a basic Sudoku game. You can switch between filling in notes
(pencil mode), or filling in answers (pen mode). To make it easier to
see where to place numbers, you can highlight all instances of a number.

Tetzle is a jigsaw puzzle game that uses tetrominoes for the pieces. Any
image can be imported and used to create puzzles with a wide range of
sizes.

Peg-E is a peg solitaire game in which you jump over pieces in order to
remove them from the board, ultimately trying to eliminate all but one.

ok ajacoutot@