import spice-protocol-0.12.2
The Simple Protocol for Independent Computing Environments (SPICE) is a
remote display system built for virtual environments which allows you to
view a computing 'desktop' environment not only on the machine where it
is running, but from anywhere on the Internet and from a wide variety of
machine architectures.

This package provides headers defining protocols.

ok aja@ gsoares@