import py-rdflib-3.2.2
RDFLib is a Python library for working with RDF, a simple yet powerful
language for representing information.  The library contains parsers and
serializers for RDF/XML, N3, NTriples, Turtle, TriX and RDFa.  The
library presents a Graph interface which can be backed by any one of a
number of Store implementations.  This core RDFLib package includes
store implementations for in memory storage and persistent storage on
top of the Berkeley DB.

ok aja@