Import foma-0.9.16.
Foma is a compiler, programming language, and C library for constructing
finite-state automata and transducers for various uses. It has specific
support for many natural language processing applications such as
producing morphological analyzers. Although NLP applications are
probably the main use of foma, it is sufficiently generic to use for a
large number of purposes.

ok jeremy@ pirofti@