Import libattica, KDE 4 prerequisite.
Attica is a Qt library that implements the Open Collaboration
Services API version 1.4.

It grants easy access to the services such as querying information
about persons and contents.  The library is used in KNewStuff3 as
content provider.  In order to integrate with KDE's Plasma Desktop,
a platform plugin exists in kdebase.

ok landry@