Add a "convenient dependency" on gnome-icon-theme-symbolic to prevent hunting what needs what: => some apps need more icons than the stock ones => some apps require symbolic icons with no fallback => some apps needs the svg version of the icon
It makes no sense to hunt all the ports tree source files to check what
icon are used exactly. That means than from now on we are _certain_ not
to miss any icons in gtk applications that are run outside of a full
blown Desktop environment.

no objection from landry@ naddy@ espie@