KolorLines is a simple but highly addictive, one player game for KDE. KolorLines has been inspired by well known game of Color Lines, written for DOS, by Olga Demina, Igor Ivkin and Gennady Denisov back in 1992.
The goal of KolorLines is quite plain. The player has to move the
colored balls around the game board, gathering them into the lines of
the same color by five. Once the line is complete it is removed from the
board, therefore freeing precious space. In the same time the new balls
keep arriving by three after each move, filling up the game board.

KolorLines cannot be won, and is played against the high score
exclusively. The game ends once the whole game board is filled up
with balls.