import abcl-1.1.1
Armed Bear Common Lisp (ABCL) is a full implementation of the Common
Lisp language featuring both an interpreter and a compiler, running in
the JVM. Originally started to be a scripting language for the J editor,
it now supports JSR-223 (Java scripting API): it can be a scripting
engine in any Java application. Additionally, it can be used to
implement (parts of) the application using Java to Lisp integration
APIs.

from timo.myyra@wickedbsd.net
ok aja@