Import ports/mail/lumail. ok landry@
lumail is a modern console-base e-mail client. It operates exclusively
on Maildir hierarchies, there is no support for POP3 or IMAP. luamail
has built in Lua scripting support. Its configuration is done in Lua
and there are a number of useful primitives you can use to control its
behavior and functionality.