Update to harfbuzz-0.9.18.
There is a major change in the port: the library has been split into 2
subpackages:
-main
-icu which provides ICU support
This allows not to force everything to link against libicu.

Breakage *is* expected but a bulk will be run shortly to cope with that
and a hackathon is a good time to do that.
Webkit is being worked on by myself.