Import hs-lens-3.7.6:
This package comes \"Batteries Included\" with many useful lenses
for the types commonly used from the Haskell Platform, and with
tools for automatically generating lenses and isomorphisms for
user-supplied data types.
The combinators in @Control.Lens@ provide a highly generic toolbox
for composing families of getters, folds, isomorphisms, traversals,
setters and lenses and their indexed variants.

ok sthen@

(will be hooked to the build during the ghc update)