Initial import of p5-File-Find-Object 0.2.7
File::Find::Object does the same job as File::Find but works like
an object and with an iterator.  As File::Find is not object oriented,
one cannot perform multiple searches in the same application.  The
second problem of File::Find is its file processing: after starting
its main loop, one cannot easily wait for another event and so get
the next result.