Import salt 0.16.3.
Salt is a new approach to infrastructure management. Easy enough to get
running in minutes, scalable enough to manage tens of thousands of
servers, and fast enough to communicate with them in seconds.

Salt delivers a dynamic communication bus for infrastructures that can
be used for orchestration, remote execution, configuration management
and much more.

ok ajacoutot@