Import py-argh 0.21.2.
Did you ever say "argh" trying to remember the details of optparse or
argparse API? If yes, this package may be useful for you. It provides a
very simple wrapper for argparse with support for hierarchical commands
that can be bound to modules or classes. Argparse can do it; argh makes
it easy.

ok ajacoutot@