import ports/sysutils/tray-app, ok dcoppa@ (plus a wantlib sync since then as I missed something).
Tray-app is a set of small gtk programs that reside in the system tray
(aka notification area).

eject ("safely remove hardware") allows you to choose a device from a
list, unmount filesystems and eject the device. This would normally be
run using sudo to elevate privileges.

battery is a battery charge level meter.

sound is an audio mixer.