Import vms-empire-1.10.
Empire is a war game played between you and the computer. The world
on which the game takes place is a square rectangle containing cities,
land, and water. Cities are used to build armies, planes, and ships
which can move across the world destroying enemy pieces, exploring, and
capturing more cities. The objective of the game is to destroy all the
enemy pieces, and capture all the cities. The computer plays by the same
rules you do.

This game was ancestral to all later expand/explore/exploit/exterminate
games, including Civilization and Master of Orion.

ok benoit@ juanfra@