Import link-grammar, a dependency for upcoming KDE 4.11.3 PIM.
The Link Grammar Parser is a syntactic parser of English, Russian,
Arabic and Persian (and other languages as well), based on link
grammar, an original theory of English syntax. Given a sentence,
the system assigns to it a syntactic structure, which consists of
a set of labelled links connecting pairs of words. The parser also
produces a "constituent" (Penn tree-bank style phrase tree)
representation of a sentence (showing noun phrases, verb phrases,
etc.). The RelEx extension provides dependency-parse output.

okay landry@, espie@ (for the BUILD_PACKAGES part)