Install the fontconfig configuration: * Sync font directory list with base system. * Put the user font cache directory under: XDG_CACHE_HOME/.cache/fontconfig-compat_linux to make it clear it is used by the Linux compatibility layer. (fontconfig in Fedora 10 does not support the "xdg" prefix for cachedir)
This prevents creating the fontconfig cachedir directly in the user home
directory.

WARNING: due to a symlink replaced by a directory, this package can not
be properly updated, so a pkg_delete+pkg_add will be needed; since we
are riding the time_t switch, this will not be an issue for regular
release update since all pkg must be removed anyway.