Remove the -browser subpackage.
It's broken at runtime anyway since the switch from xul 1.8 to 1.9 three
years ago, is now the last user of xul 1.9, and noone is probably using
an eclipse from 2007 to do "serious" java development. Lots of cursing
involved to find the correct magic to properly disable the building of
swt-mozilla-gtk.

Discussed with kurt@ (MAINTAINER)