Kopete is an instant messenger supporting AIM, Bonjour, Gadu-Gadu, GroupWise, ICQ, Jabber (XMPP, Google Talk, Facebook, ...), Meanwhile, QQ, Skype, Windows Live Messenger, WinPopup, Yahoo and more. It is designed to be a flexible and extensible multi-protocol system suitable for personal and enterprise use.
The goal of Kopete is to provide users with a single easy-to-use
way to access all of their instant messaging systems. The interface
puts people first, and is integrated with the system address book
to let you access your contacts from other KDE applications. Instant
messaging can be intrusive, but Kopete's notification system can
be tuned so that only important contacts interrupt you.

Kopete also features tools to enhance your instant messaging
experience, such as message encryption, archiving, and many other
fun and useful effects.

This is ex-kdenetwork member. Importing as a part of KDE 4.11.4 update.
This port won't work until update is finished, thus not hooking it up.