import devel/kyua-cli, from Sergey Bronnikov, ok shadchin@ pre-lock, only minor changes since.
Kyua (pronounced Q.A.) is a testing framework for both developers and
users. Kyua is different from most other testing frameworks in that it
puts the end user experience before anything else. There are multiple
reasons for users to run the tests themselves, and Kyua ensures that
they can do so in the most convenient way.