Import flac123.
flac123 is a command line tool for playing audio files stored in
the lossless FLAC format. It fills a gap in the existing FLAC tools
which provide a standalone tool to encode and decode files, a tool
to manipulate metadata, but no command line tool for playback.

From: Donovan Watteau <tsoomi@gmail.com>