Import games/vegastrike, ok bentley@.
Vega Strike is an OpenSource 3d Space Simulator. Work is proceeding
on both the Vega Strike game engine and a Vega Strike Universe game
data set. Both engine and data are in a playable state, but still
under development. Each release serves as a beta for both engine
and data.

The Vega Strike engine serves as the foundation for the Privateer
Gemini Gold, Privateer Remake, Vega Trek, and Pi Armada projects.