Update to firefox 29.0, with the brand new Australis UI.
- See http://www.mozilla.org/en-US/firefox/29.0/releasenotes/
- Fixes MFSA 2014-34->47
- add spellchecker.dictionary_path as a default pref in all-openbsd.js,
pointing to ${LOCALBASE}/share/mozilla-dicts/ so that it looks for
systemwide dictionaries.
- Remove now useless patch-extensions_spellcheck_hunspell_src_mozHunspell_cpp
- Add patch for #990196, tries to use INSTALL_SCRIPT during build which
  badly fails as user.

Note that starting with Fx 29, the sync protocol v1.1 used by
www/mozilla-sync still works but is not advertised anymore, and will be
phased out by v1.5 in Fx 31. www/mozilla-sync port will be updated in the
meantime, but users will have to reconfigure all their instances...