Remove databases/ruby-datamapper and subdirs
These used version 1.0.2 which is about four years out of date, and
development stopped with 1.2.0 a couple years ago.  Nothing in the
tree depends on these anymore.

No quirks addition for this removal, as there are 11 subdirs each
buildable with 6 different ruby interpreters (with different
package names per interpreter), and this doesn't warrant the
addition of 66 quirks.

OK ajacoutot@