Import dunelegacy, an engine for DuneII game.
Lead one of three interplanetary houses, Atreides, Harkonnen or Ordos,
in an attempt to harvest the largest amount of spice from the sand
dunes. Exchange your spice stockpiles for credits through refinement and
build an army capable of thwarting attempts of the other houses to stop
your harvesting!

Dune Legacy is an effort by a handful of developers to revitalize one of
the first real-time strategy games. The original game was the basis for
the hugely successful Command and Conquer series, and the gameplay has
been replicated an extended to a wide variety of storylines and series.

help and OK edd@