import portscout-0.8.1.1 (0.8.1 with OpenBSD support)
Portscout is a tool which looks for new versions of software in the
FreeBSD ports tree, and potentially other software repositories.

Various factors make this task a bit more difficult than it might
initially seem. In particular, the array of weird and wonderful
versioning schemes software vendors manage to come up with.

A list of results can be found here: http://portscout.jasper.la/

ok ajacoutot@