Import p5-Modern-Perl 1.20140107.
Modern Perl programs use several modules to enable additional features of Perl
and of the CPAN. Instead of copying and pasting all of these use lines, instead
write only one:

    use Modern::Perl;

This enables the strict and warnings pragmas, as well as all of the features
available in Perl 5.10. It also enables C3 method resolution order as
documented in perldoc mro and loads IO::File and IO::Handle so that you may
call methods on filehandles. In the future, it may include additional core
modules and pragmas.

From MAINTAINER edward # rdtan / net, ok sthen@