Import synfigstudio-0.64.1.
synfig is a vector based 2D animation package. It is designed to be
capable of producing feature-film quality animation. It eliminates the
need for tweening, preventing the need to hand-draw each frame. synfig
features spatial and temporal resolution independence (sharp and smooth
at any resolution or framerate), high dynamic range images, and a
flexible plugin system.

synfigstudio is the animation studio for synfig and provides the GUI
interface to create synfig animations which are saved in synfig .sif
or .sifz format.

ok jca@