Import a KDE4 version, 4.0.9, of KRename.
KRename is a powerful batch renamer for KDE.  It allows you to easily
rename hundreds or even more files in one go.  The filenames can be
created by parts of the original filename, numbering the files or
accessing hundreds of informations about the file, like creation date or
Exif informations of an image, thanks to KDE file plugins.

okay landry@