import california-0.2.0
California is a new calendar application for GNOME 3 designed to let you
work with your calendars quickly and easily.  It provides month and week
views of your schedule and easy ways to add and update events, including
a Quick Add feature which allows you to type the event details and have
them added to your calendar without filling out a form.

ok aja@