print/texlive: Remove a load of internal libraries, fix readme.
This started as a fix (arm, ppc) for the internal icu in texlive. It turns
out a custom icu is not needed anymore, so we can use the icu in the ports
tree. This makes handling of header files much simpler since there is no
longer the possibility of picking up the wrong icu headers. Should also fix
PPC and ARM builds of texlive.

Also noticed a few other libraries being built into texlive. Fixed.

Finally, readme was not being installed. Fixed.

Reported by landry@

OK dcoppa@, landry@, and with input and testing from kili@.