Import megatools-1.9.94.
Megatools is a collection of programs for accessing Mega service from a
command line of your desktop or server.

Megatools allow you to copy individual files as well as entire directory
trees to and from the cloud. You can also perform streaming downloads
for example to preview videos and audio files, without needing to
download the entire file.

Megatools are robust and optimized for fast operation - as fast as Mega
servers allow. Memory requirements and CPU utilization are kept at
minimum.

ok juanfra@