Import pavucontrol 2.0.
PulseAudio Volume Control (pavucontrol) is a simple GTK based volume
control tool ("mixer") for the PulseAudio sound server. In contrast
to classic mixer tools this one allows you to control both the volume
of hardware devices and of each playback stream separately.

Default to build with gtk2, there's a gtk3 flavor that doesnt work yet
but ajacoutot@ has a fix for it.

ok ajacoutot@